@@ -21,7 +21,7 @@ class ExceptionEnricherProcessorTest extends TestCase
      */
     public function testEmptyProcessor()
     {
-        $exceptionEnricherProcessor = new ExceptionEnricherProcessor(null, null, null);
+        $exceptionEnricherProcessor = new ExceptionEnricherProcessor(null, null);
         $record = $exceptionEnricherProcessor($this->createRecord());
 
         $this->assertArrayNotHasKey('request_uri', $record['extra']);
@@ -30,7 +30,7 @@ class ExceptionEnricherProcessorTest extends TestCase
         $this->assertArrayNotHasKey('request_ip', $record['extra']);
         $this->assertArrayNotHasKey('session_id', $record['extra']);
         $this->assertArrayNotHasKey('username', $record['extra']);
-        $this->assertTrue(empty($record['extra']));
+        $this->assertEmpty($record['extra']);
     }
 
     /**
@@ -44,20 +44,21 @@ class ExceptionEnricherProcessorTest extends TestCase
         $request->getMethod()->willReturn('GET')->shouldBeCalled();
         $request->getRequestUri()->willReturn('/testroute/')->shouldBeCalled();
 
-        $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal())->shouldBeCalled();
-        $requestStack->getMasterRequest()->willReturn($request->reveal())->shouldBeCalled();
-
         $session = $this->prophesize(SessionInterface::class);
         $session->getId()->willReturn('39d9f31fb12441428031e26d2f83ab6e')->shouldBeCalled();
 
+        $requestStack = $this->prophesize(RequestStack::class);
+        $requestStack->getCurrentRequest()->willReturn($request->reveal())->shouldBeCalled();
+        $requestStack->getMainRequest()->willReturn($request->reveal())->shouldBeCalled();
+        $requestStack->getSession()->willReturn($session)->shouldBeCalled();
+
         $token = $this->prophesize(TokenInterface::class);
-        $token->getUsername()->willReturn('testuser')->shouldBeCalled();
+        $token->getUserIdentifier()->willReturn('testuser')->shouldBeCalled();
 
         $tokenStorage = $this->prophesize(TokenStorageInterface::class);
         $tokenStorage->getToken()->willReturn($token->reveal())->shouldBeCalled();
 
-        $exceptionEnricherProcessor = new ExceptionEnricherProcessor($requestStack->reveal(), $session->reveal(), $tokenStorage->reveal());
+        $exceptionEnricherProcessor = new ExceptionEnricherProcessor($requestStack->reveal(), $tokenStorage->reveal());
         $record = $exceptionEnricherProcessor($this->createRecord());
 
         $this->assertArrayHasKey('request_uri', $record['extra']);
